Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.10
aiofiles>=24.0.0

sqlalchemy>=2.0.30
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
Mako>=1.3.0

pydantic>=2.9.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
bcrypt>=4.2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

limits>=3.13.0
redis>=5.0.0
gunicorn>=23.0.0

supabase>=2.8.0